azureml_sample_project/requirements.txt
azure-ai-ml
azure-identity
pandas
requests
pyarrow
scikit-learn
joblib
